Create separate crate for flight_proto just containing the compiled code
