+++ rust/flight/src/bin/main.rs
@@ -15,8 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use arrow_flight::server::FlightServiceServer;
-use flight::*;
+use flight_proto::arrow_flight::server::FlightServiceServer;
+use flight_proto::FlightServiceImpl;
 use tonic::transport::Server;
 
 #[tokio::main]
