+++ rust/rust-toolchain
@@ -1,0 +1,1 @@
+nightly-2019-11-14
